+++ xoeuf/tests/test_xoeuf_api/__openerp__.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# ---------------------------------------------------------------------
+# Copyright (c) Merchise Autrement [~º/~] and Contributors
+# All rights reserved.
+#
+# This is free software; you can do what the LICENCE file allows you to.
+#
+
+{
+    'name': 'test_api',
+    'author': 'Merchise Autrement [~º/~] and Contributors',
+    'description': 'Test for our API functions',
+    'depends': ['base', ],
+    'installable': True,
+    'auto_install': False,
+}